Session shutdown cleans up project dangling images
