+++ services/orchest-api/app/app/apis/namespace_environment_images.py
@@ -14,13 +14,13 @@ from app.apis.namespace_jobs import AbortJob
 from app.apis.namespace_runs import AbortPipelineRun
 from app.apis.namespace_sessions import StopInteractiveSession
 from app.connections import db, docker_client
+from app.core import docker_utils
 from app.utils import (
     interactive_runs_using_environment,
     interactive_sessions_using_environment,
     is_environment_in_use,
     jobs_using_environment,
     register_schema,
-    remove_if_dangling,
 )
 
 api = Namespace("environment-images", description="Managing environment images")
@@ -74,59 +74,10 @@ class ProjectEnvironmentDanglingImages(Resource):
         tag-less and which are not referenced by any run
         or job which are pending or running."""
 
-        delete_project_environment_dangling_images(project_uuid, environment_uuid)
+        docker_utils.delete_project_environment_dangling_images(
+            project_uuid, environment_uuid
+        )
         return {"message": "Successfully removed dangling images."}, 200
-
-
-def delete_project_dangling_images(project_uuid):
-    """Removes dangling images related to a project.
-
-    Dangling images are images that have been left nameless and
-    tag-less and which are not referenced by any run
-    or job which are pending or running.
-
-    Args:
-        project_uuid:
-    """
-    # Look only through runs belonging to the project.
-    filters = {
-        "label": [
-            "_orchest_env_build_is_intermediate=0",
-            f"_orchest_project_uuid={project_uuid}",
-        ]
-    }
-
-    project_images = docker_images_list_safe(docker_client, filters=filters)
-
-    for docker_img in project_images:
-        remove_if_dangling(docker_img)
-
-
-def delete_project_environment_dangling_images(project_uuid, environment_uuid):
-    """Removes dangling images related to an environment.
-
-    Dangling images are images that have been left nameless and
-    tag-less and which are not referenced by any run
-    or job which are pending or running.
-
-    Args:
-        project_uuid:
-        environment_uuid:
-    """
-    # Look only through runs belonging to the project consider only
-    # docker ids related to the environment_uuid.
-    filters = {
-        "label": [
-            "_orchest_env_build_is_intermediate=0",
-            f"_orchest_project_uuid={project_uuid}",
-            f"_orchest_environment_uuid={environment_uuid}",
-        ]
-    }
-
-    project_env_images = docker_images_list_safe(docker_client, filters=filters)
-
-    for docker_img in project_env_images:
-        remove_if_dangling(docker_img)
 
 
 class DeleteProjectEnvironmentImages(TwoPhaseFunction):
@@ -188,7 +139,9 @@ class DeleteImage(TwoPhaseFunction):
             project_uuid=project_uuid, environment_uuid=environment_uuid
         )
 
-        delete_project_environment_dangling_images(project_uuid, environment_uuid)
+        docker_utils.delete_project_environment_dangling_images(
+            project_uuid, environment_uuid
+        )
         if os.fork() == 0:
             # Try with repeat because there might be a race condition
             # where the aborted runs are still using the image. Moreover
