+++ services/orchest-api/app/app/core/docker_utils.py
@@ -5,6 +5,7 @@ import time
 import docker
 
 from _orchest.internals.utils import docker_images_list_safe
+from app import utils
 from app.connections import docker_client
 
 __DOCKERFILE_RESERVED_FLAG = "_ORCHEST_RESERVED_FLAG_"
@@ -182,3 +183,54 @@ def cleanup_docker_artifacts(filters):
                 )
                 break
             time.sleep(1)
+
+
+def delete_project_environment_dangling_images(project_uuid, environment_uuid):
+    """Removes dangling images related to an environment.
+
+    Dangling images are images that have been left nameless and
+    tag-less and which are not referenced by any run
+    or job which are pending or running.
+
+    Args:
+        project_uuid:
+        environment_uuid:
+    """
+    # Look only through runs belonging to the project consider only
+    # docker ids related to the environment_uuid.
+    filters = {
+        "label": [
+            "_orchest_env_build_is_intermediate=0",
+            f"_orchest_project_uuid={project_uuid}",
+            f"_orchest_environment_uuid={environment_uuid}",
+        ]
+    }
+
+    project_env_images = docker_images_list_safe(docker_client, filters=filters)
+
+    for docker_img in project_env_images:
+        utils.remove_if_dangling(docker_img)
+
+
+def delete_project_dangling_images(project_uuid):
+    """Removes dangling images related to a project.
+
+    Dangling images are images that have been left nameless and
+    tag-less and which are not referenced by any run
+    or job which are pending or running.
+
+    Args:
+        project_uuid:
+    """
+    # Look only through runs belonging to the project.
+    filters = {
+        "label": [
+            "_orchest_env_build_is_intermediate=0",
+            f"_orchest_project_uuid={project_uuid}",
+        ]
+    }
+
+    project_images = docker_images_list_safe(docker_client, filters=filters)
+
+    for docker_img in project_images:
+        utils.remove_if_dangling(docker_img)
