+++ services/orchest-api/app/app/core/sessions.py
@@ -17,6 +17,7 @@ from flask import current_app
 
 from _orchest.internals import config as _config
 from app import errors, utils
+from app.core import docker_utils
 
 
 class SessionType(Enum):
@@ -376,6 +377,7 @@ class Session:
             except (requests.exceptions.HTTPError, NotFound, APIError):
                 pass
 
+        docker_utils.delete_project_dangling_images(project_uuid)
         return
